First commit
This holds a lot, and it contains the initial .urdf file handwritten using urdf boxes, as well as a newer xacro macro generated robot description using more accurate sensor descriptions and meshes. It includes a launch file which starts up the simulation and can be configured by the user as well.